flasksocket/requirements.txt
Flask
Flask-SocketIO==5.3.4
gevent-websocket==0.10.1
gunicorn
redis